Refine the model ready_set produced, and write plain floats to YAML
phenix.refine refuses to start when it cannot type an atom, and a deposited
model gives it two reasons: ligands absent from the monomer library (7LVC's
NADP, 73 atoms) and a stray hydrogen with no restraint definition (one on
its folate). ready_set resolves both, writing restraint CIFs and an updated
model -- but only the CIFs were being used, so the count fell from 73 to 1
and refinement still refused. It now refines the updated model.

Separately, combine_sweeps wrote numpy scalars into dataset.yaml, which
yaml.safe_dump refuses; the statistics are cast to float.

+++ scripts/sbgrid/combine_sweeps.py
@@ -70,9 +70,11 @@ def anscombe_stats(counts: np.memmap, masks: np.memmap, chunk: int) -> tuple:
         ans_sq += (transformed**2).sum()
     raw_mean = raw_sum / total
     ans_mean = ans_sum / total
+    # plain floats: yaml.safe_dump refuses a numpy scalar, and these come out
+    # of numpy reductions
     return (
-        [raw_mean, raw_sq / total - raw_mean**2],
-        [ans_mean, ans_sq / total - ans_mean**2],
+        [float(raw_mean), float(raw_sq / total - raw_mean**2)],
+        [float(ans_mean), float(ans_sq / total - ans_mean**2)],
     )
